Add function to modify mapping layout with a button press
Pressing the Y button at the top level menu will allow
mapping a special function to the source button.

The destination consists of one of the 4 D-pad directions who
correspond respectively to one of the mapping layouts.

While holding a button set to this function, the current
mapping layout configured will be overwritten until
the button is released.