GameCube joysticks scaling
Fixes #7
GameCube joysticks range differs from the N64. N64 has a maximum of ~84 along the axes origin
and ~71 in the diagonals. GameCube main joystick has a maximum (once dead zone & sign applied)
of ~90 and ~65 for the same. C joystick maximums are a bit lower at ~82 and ~57. N64 max value
once plot will give us a square-ish equilateral hexagon while GameCube is an equiangular &
equilateral hexagon.

This may be problematic in some games where the diagonal max speed could be lower. In other games
the joystick may feel too sensitive since the main GC joystick output values are significantly
higher along the origin.

This feature uses the opposite axis as a reference to determine dynamically the scaling value
required. The scaling values are stored in a table using the fixed point format of 1.7. Once
multiplied by the joystick value (in 8.0 format) this gives us in our hardware multiplier a
value in the fixed point 9.7 format.

The D-pad LEFT submenu can disable this function (X option) as it may be beneficial in some case
to use the previous behavior. (Default enabled)