Add bypass mode for using N64 controller as accessory slot
Fixes #10
The adaptor can now have a N64 controller attach to it on RA2 to be
used as an accessory slot. Once in this mode the adaptor will only
answer the button status commands. All other commands will be copied to
the N64 controller. Then the controller response is copied back to the
system.

Pressing D-pad Up in the adaptor mode submenu (or at boot) will activate
the function. Pressing D-pad Left or Right will either resume Rumble
emulation or display the adaptor slot as empty.

When toggling between mode, the adaptor will observe a 1 s period
where it will show the slot as empty. This will force the games to
re-identify the accessory in the slot.