Support 1 player 2 controllers control style
GoldenEye and Perfect Dark got a control style using
two controllers as a way to get dual analog sticks.

Since the adaptor is now interrupt base it's quite easy now
to answer another poll on another pin.

This duplicate the mapping of the C buttons (as analog) and R buttons into
second set of buffers for the second controller.

The second controller interrupt is only enabled if the adaptor detect
at boot that the second port is connected.