Restore RA4 transition detection in serial bridge
Somehow the published version from 2004 do not check for
transition on RA4, it rather sent TMR0 value.

Fix up as well couple Python 3 script issue.

It's purely academic but it allows recreating Micah original
experiment from 2004!

Traces done with the serial bridge on the controller bus are
in the ctrl_bus_trace folder.