Add a special two-button mapping for A+C-Down & B+C-Left
In all three Tony Hawk's games the only right side face buttons used are
the C Buttons, while in the menu only the A & B buttons are used.

Mapping the GC face buttons to C Buttons make menu navigation
annoying.

Since there is no overlap between the menu and gameplay, this
special mapping function merge on the same buttons both the A & C-Down
and B & C-Right.