Menu cleanup
Renamed all labels to make clearer at which level of the menu
they are used.

Also moved the flags that control the transition between menu level
to its own variable: atomic_flag. Any multiple bit change carried on this
variable need to be done in an atomic operation. This avoids the interrupt
placing the firmware in an invalid state.

Other flags shouldn't require any particular handling.