Pre-defined mapping layout presets
Add a new entry in the main menu via D-pad DOWN for pre-defined mapping layout presets.
Eight preset are available on flash to be copied in one of the four mapping banks in the EEPROM.

After pressing D-pad DOWN in the menu, the next input on the 4 D-pad or 4 main joystick direction
will load the presets chosen in the currently active mapping layout bank in the EEPROM.

See wiki for presets detail.