Allow in GoldenEye to shoot/switch weapon in the cut-scene via 2nd CTRL
This is a speed run trick that requires two changes in the
1 player 2 controllers feature:

1. A new special button that can be mapped via X in the special
feature menu. When pressed, this special button will mute the
buttons reported to the first controller slot (Exept Z & Start).
Without the mute the press on the 1st controller would skip the cut scene.

2. A & B button press is added to the 2nd controller slot
status. This allows to switch weapons.